requirements.txt
cffi==1.7.0
cryptography==1.4
enum34==1.1.6
github3.py==1.0.0a4
idna==2.1
ipaddress==1.0.16
ndg-httpsclient==0.4.2
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
pytz==2016.6.1
requests==2.11.1
six==1.10.0
uritemplate.py==2.0.0
